+++ app/src/main/java/com/app/sha/attar/invoice/adapter/ReportViewAdapter.java
@@ -36,9 +36,9 @@ public class ReportViewAdapter extends RecyclerView.Adapter<ReportViewHolder> {
 
         holder.name.setText(displayList.get(index).getName());
         holder.quantity.setText(String.valueOf(displayList.get(index).getQuantity()));
-        holder.actual_amt.setText("Rs. "+String.valueOf((int)displayList.get(index).getActualPrice()));
-        holder.selling_amount.setText("Rs. "+String.valueOf((int)displayList.get(index).getSoldPrice()));
-        holder.profit.setText("Rs. "+String.valueOf((int)displayList.get(index).getProfit()));
+        holder.actual_amt.setText("Rs. "+String.valueOf((double)displayList.get(index).getActualPrice()));
+        holder.selling_amount.setText("Rs. "+String.valueOf((double)displayList.get(index).getSoldPrice()));
+        holder.profit.setText("Rs. "+String.valueOf((double)displayList.get(index).getProfit()));
 
     }
